test(jetstream): revive misnamed kube cert auth test
cert_tests.go never matched go test's _test.go convention, so
TestFetchCertAuth compiled into the production package (dragging
testing/httptest along) and never ran. Renamed to cert_test.go and
updated the request to the current extractCerts contract: cert and
certKey as form fields rather than a raw cert:key body. Replaced
t.Fail() with t.Fatal on setup errors — the old flow continued past
a nil result and segfaulted once the test actually executed.
